fix(core): address adversarial review findings
VSDD Phase 3/4 fixes from adversarial code roast:
- state_machine: carry calibration_valid/last_disarm_ms in sm_result_t;
  add sm_apply() helper; cutoff in Calibrate transitions to Diag;
  add REPORT_LOWBATT on critical; invalid mode returns MOTORS_OFF
- calibration: use int64_t for sum/variance to prevent overflow
- pid: guard derivative term with dt > 1e-6f to prevent Inf/NaN
- mixer: add isfinite() guard in clamp to prevent UB on NaN input
- battery: allow recovery from CUTOFF when voltage rises
- microblue_parser: reject oversized IDs instead of truncating
- input_mapper: clamp joystick raw input to [0, JOYSTICK_MAX]

+++ lib/core/calibration.c
@@ -14,8 +14,8 @@ calib_result_t calib_compute(const int16_t (*samples)[CALIB_AXES],
 		return result;
 	}
 
-	/* Compute mean for each axis */
-	int32_t sum[CALIB_AXES] = {0};
+	/* Compute mean for each axis (int64 to avoid overflow) */
+	int64_t sum[CALIB_AXES] = {0};
 	for (uint16_t i = 0; i < sample_count; i++) {
 		for (int a = 0; a < CALIB_AXES; a++) {
 			sum[a] += samples[i][a];
@@ -28,12 +28,12 @@ calib_result_t calib_compute(const int16_t (*samples)[CALIB_AXES],
 	/* Compute variance for each axis and check threshold */
 	result.ok = true;
 	for (int a = 0; a < CALIB_AXES; a++) {
-		int32_t variance_sum = 0;
+		int64_t variance_sum = 0;
 		for (uint16_t i = 0; i < sample_count; i++) {
 			int32_t diff = samples[i][a] - result.offsets[a];
-			variance_sum += diff * diff;
+			variance_sum += (int64_t)diff * diff;
 		}
-		int32_t variance = variance_sum / sample_count;
+		int64_t variance = variance_sum / sample_count;
 		if (variance > CALIB_VARIANCE_THRESHOLD) {
 			result.ok = false;
 			break;
